pyproject.toml
[tool.poetry]
name = "itemsmanager"
version = "0.1.0"
description = ""
authors = ["'dmytro-ustynov' <'ustynov.dmitriy@gmail.com'>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
python-decouple = "^3.8"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
python-multipart = "^0.0.9"
openpyxl = "^3.1.4"
pymongo = "^4.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
